Dz4 for Java - Fix

+++ Go_Start/Java/src/dz4/Task4_1.java
@@ -1,11 +1,11 @@
 package dz4;
 
 public class Task4_1 {
-// Твой Вариант, как убрать нолики не разобрался...
+// Твой Вариант
 
     public static void main(String[] args) {
         for (int i = 1; i <= 20; i++) {
-            System.out.println(Math.pow(2, i));
+            System.out.println((int)Math.pow(2, i));
         }
     }
 }
